requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==1.2.0
Django==4.1.7
django-embed-video==1.4.8
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
